Remove base_path as a default
The base_path variable was defaulted to /mnt/volume in the inventory.sample file.  This led to some issues for novice users as the path doesn't exist on their host and the service fails to start.  I believe the default path (/home/polkadot/.local/...) should be used in all cases and /mnt/volume be used by exception.